Added Settings screen
Implemented Settings screen, now available to user with functionality of:
 - Showing the user the heater target temperature
 - Allowing user to use up/down buttons to change heater target temperature
 - Heater Mode can be changed

Bug fixes:
 - Changed HeaterTargetTemperature to use float not integer.
 - Changed the struct Limits to use floats not integer.
 - Multiple ScreenData "Hard coded" macro instead of specific instance named macro.

Changed struct Limits to Limits_Float, being more precise in case integer version needed later.

Added heater target temperature limits.

Removed/Added SPACING_X_PIXELS macros.

+++ IoT_LCD_Screen.ino
@@ -18,11 +18,9 @@ uint32_t PreviousLogicTime = 0;
 uint8_t ChangeMenuFlag = 0;
 uint32_t CurrentTime = 0;
 // Data screen
-uint8_t HeaterTargetTemperature = DEFAULT_VAL_HEATER_TARGET_TEMPERATURE;
+float HeaterTargetTemperature = DEFAULT_VAL_HEATER_TARGET_TEMPERATURE;
 uint8_t HeaterOnOffState = 0;
 uint8_t HeaterMode = 0; // Default Auto
-// Settings screen
-// TODO: This should be
 // Keypad screen
 uint16_t UserNumber = 0;
